refactor(workflow): fold wardian-workflow into wardian-core::workflow
