+++ crates/wardian-cli/Cargo.toml
@@ -18,7 +18,6 @@ serde = { workspace = true }
 serde_json = { workspace = true, features = ["preserve_order"] }
 tokio = { version = "1.49.0", features = ["io-util", "net", "rt", "time"] }
 wardian-core = { path = "../wardian-core" }
-wardian-workflow = { path = "../wardian-workflow" }
 
 [dev-dependencies]
 tempfile = { workspace = true }
