+++ crates/wardian-core/Cargo.toml
@@ -13,10 +13,10 @@ dirs = { workspace = true }
 once_cell = { workspace = true }
 rusqlite = { workspace = true }
 serde = { workspace = true }
-serde_json = { workspace = true }
+serde_json = { workspace = true, features = ["preserve_order"] }
+serde_norway = "0.9"
 thiserror = { workspace = true }
 uuid = { workspace = true }
 
 [dev-dependencies]
 tempfile = { workspace = true }
-
