+++ crates/wardian-engine/Cargo.toml
@@ -8,7 +8,7 @@ repository.workspace = true
 homepage.workspace = true
 
 [dependencies]
-wardian-workflow = { path = "../wardian-workflow" }
+wardian-core = { path = "../wardian-core" }
 serde = { workspace = true }
 serde_json = { workspace = true, features = ["preserve_order"] }
 thiserror = { workspace = true }
