+++ crates/wardian-engine/src/graph.rs
@@ -1,4 +1,4 @@
-use wardian_workflow::Blueprint;
+use wardian_core::workflow::Blueprint;
 
 /// Precomputed adjacency over a blueprint's edges, addressed by edge index so
 /// `RunState` can track delivered/skipped edges compactly.
@@ -72,7 +72,7 @@ impl<'a> Graph<'a> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use wardian_workflow::{Blueprint, Edge, Node};
+    use wardian_core::workflow::{Blueprint, Edge, Node};
 
     fn n(id: &str, ty: &str, parent: Option<&str>) -> Node {
         Node {
