--- crates/wardian-workflow/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "wardian-workflow"
-version.workspace = true
-edition.workspace = true
-authors.workspace = true
-license.workspace = true
-repository.workspace = true
-homepage.workspace = true
-
-[dependencies]
-serde = { workspace = true }
-serde_json = { workspace = true, features = ["preserve_order"] }
-serde_norway = "0.9"
-thiserror = { workspace = true }
-once_cell = { workspace = true }
-
-[dev-dependencies]
-tempfile = { workspace = true }
